svi_pipeline/requirements.txt
openai>=1.30.0
openpyxl>=3.0.10
numpy>=1.23.0
pandas>=1.5.3
scikit-learn>=1.1.0
matplotlib>=3.6.0
seaborn>=0.12.0
opencv-python>=4.7.0
xgboost>=1.7.0
lightgbm>=3.3.5
scipy>=1.9.0
tqdm>=4.64.0
joblib>=1.2.0
PyYAML>=6.0.0
